Clang-format to the code...

+++ src/IdentityManager/identitymanager_db_roles.cpp
@@ -1,8 +1,7 @@
 #include "identitymanager_db.h"
-#include <Mantids30/Threads/lock_shared.h>
 #include <Mantids30/Memory/a_string.h>
 #include <Mantids30/Memory/a_uint64.h>
-
+#include <Mantids30/Threads/lock_shared.h>
 
 using namespace Mantids30::Memory;
 using namespace Mantids30::Database;
@@ -12,27 +11,19 @@ bool IdentityManager_DB::Roles_DB::addRole(const std::string &roleName, const st
 {
     Threads::Sync::Lock_RW lock(_parent->m_mutex);
     return _parent->m_sqlConnector->query("INSERT INTO iam_roles (`roleName`,`roleDescription`) VALUES(:roleName,:roleDescription);",
-                                          {
-                                              {":roleName", MAKE_VAR(STRING,roleName)},
-                                              {":roleDescription",MAKE_VAR(STRING,roleDescription)}
-                                          });
+                                          {{":roleName", MAKE_VAR(STRING, roleName)}, {":roleDescription", MAKE_VAR(STRING, roleDescription)}});
 }
 
 bool IdentityManager_DB::Roles_DB::removeRole(const std::string &roleName)
 {
     Threads::Sync::Lock_RW lock(_parent->m_mutex);
-    return _parent->m_sqlConnector->query("DELETE FROM iam_roles WHERE `roleName`=:roleName;",
-                               {
-                                   {":roleName",MAKE_VAR(STRING,roleName)}
-                               });
+    return _parent->m_sqlConnector->query("DELETE FROM iam_roles WHERE `roleName`=:roleName;", {{":roleName", MAKE_VAR(STRING, roleName)}});
 }
 
 bool IdentityManager_DB::Roles_DB::doesRoleExist(const std::string &roleName)
 {
     Threads::Sync::Lock_RD lock(_parent->m_mutex);
-    std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `roleName` FROM iam_roles WHERE `roleName`=:roleName;",
-                                          {{":roleName",MAKE_VAR(STRING,roleName)}},
-                                          {});
+    std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `roleName` FROM iam_roles WHERE `roleName`=:roleName;", {{":roleName", MAKE_VAR(STRING, roleName)}}, {});
     return (i->getResultsOK()) && i->query->step();
 }
 
@@ -40,23 +31,19 @@ bool IdentityManager_DB::Roles_DB::addAccountToRole(const std::string &roleName,
 {
     Threads::Sync::Lock_RW lock(_parent->m_mutex);
     return _parent->m_sqlConnector->query("INSERT INTO iam_rolesAccounts (`f_roleName`,`f_accountName`) VALUES(:roleName,:accountName);",
-                               {
-                                   {":roleName",MAKE_VAR(STRING,roleName)},
-                                   {":accountName",MAKE_VAR(STRING,accountName)}
-                               });
+                                          {{":roleName", MAKE_VAR(STRING, roleName)}, {":accountName", MAKE_VAR(STRING, accountName)}});
 }
 
 bool IdentityManager_DB::Roles_DB::removeAccountFromRole(const std::string &roleName, const std::string &accountName, bool lock)
 {
     bool ret = false;
-    if (lock) _parent->m_mutex.lock();
+    if (lock)
+        _parent->m_mutex.lock();
     ret = _parent->m_sqlConnector->query("DELETE FROM iam_rolesAccounts WHERE `f_roleName`=:roleName AND `f_accountName`=:accountName;",
-                              {
-                                  {":roleName",MAKE_VAR(STRING,roleName)},
-                                  {":accountName",MAKE_VAR(STRING,accountName)}
-                              });
-    
-    if (lock) _parent->m_mutex.unlock();
+                                         {{":roleName", MAKE_VAR(STRING, roleName)}, {":accountName", MAKE_VAR(STRING, accountName)}});
+
+    if (lock)
+        _parent->m_mutex.unlock();
     return ret;
 }
 
@@ -64,10 +51,7 @@ bool IdentityManager_DB::Roles_DB::updateRoleDescription(const std::string &role
 {
     Threads::Sync::Lock_RW lock(_parent->m_mutex);
     return _parent->m_sqlConnector->query("UPDATE iam_roles SET `roleDescription`=:roleDescription WHERE `roleName`=:roleName;",
-                               {
-                                   {":roleName",MAKE_VAR(STRING,roleName)},
-                                   {":roleDescription",MAKE_VAR(STRING,roleDescription)}
-                               });
+                                          {{":roleName", MAKE_VAR(STRING, roleName)}, {":roleDescription", MAKE_VAR(STRING, roleDescription)}});
 }
 
 std::string IdentityManager_DB::Roles_DB::getRoleDescription(const std::string &roleName)
@@ -75,8 +59,7 @@ std::string IdentityManager_DB::Roles_DB::getRoleDescription(const std::string &
     Threads::Sync::Lock_RD lock(_parent->m_mutex);
     Abstract::STRING roleDescription;
     std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `roleDescription` FROM iam_roles WHERE `roleName`=:roleName LIMIT 1;",
-                                          {{":roleName",MAKE_VAR(STRING,roleName)}},
-                                          { &roleDescription });
+                                                                                      {{":roleName", MAKE_VAR(STRING, roleName)}}, {&roleDescription});
     if (i->getResultsOK() && i->query->step())
     {
         return roleDescription.getValue();
@@ -90,9 +73,7 @@ std::set<std::string> IdentityManager_DB::Roles_DB::getRolesList()
     Threads::Sync::Lock_RD lock(_parent->m_mutex);
 
     Abstract::STRING roleName;
-    std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `roleName` FROM iam_roles;",
-                                          {},
-                                          { &roleName });
+    std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `roleName` FROM iam_roles;", {}, {&roleName});
     while (i->getResultsOK() && i->query->step())
     {
         ret.insert(roleName.getValue());
@@ -100,52 +81,50 @@ std::set<std::string> IdentityManager_DB::Roles_DB::getRolesList()
     return ret;
 }
 
-
 std::set<std::string> IdentityManager_DB::Roles_DB::getRoleAccounts(const std::string &roleName, bool lock)
 {
     std::set<std::string> ret;
-    if (lock) _parent->m_mutex.lockShared();
+    if (lock)
+        _parent->m_mutex.lockShared();
 
     Abstract::STRING accountName;
     std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect("SELECT `f_accountName` FROM iam_rolesAccounts WHERE `f_roleName`=:roleName;",
-                                          { {":roleName",MAKE_VAR(STRING,roleName)} },
-                                          { &accountName });
+                                                                                      {{":roleName", MAKE_VAR(STRING, roleName)}}, {&accountName});
     while (i->getResultsOK() && i->query->step())
     {
         ret.insert(accountName.getValue());
     }
-    
-    if (lock) _parent->m_mutex.unlockShared();
+
+    if (lock)
+        _parent->m_mutex.unlockShared();
     return ret;
 }
 
-std::list<RoleDetails> IdentityManager_DB::Roles_DB::searchRoles(std::string sSearchWords, uint64_t limit, uint64_t offset)
+std::list<RoleDetails> IdentityManager_DB::Roles_DB::searchRoles(std::string sSearchWords, size_t limit, size_t offset)
 {
     std::list<RoleDetails> ret;
     Threads::Sync::Lock_RD lock(_parent->m_mutex);
 
-    Abstract::STRING roleName,description;
+    Abstract::STRING roleName, description;
 
     std::string sSqlQuery = "SELECT `roleName`,`roleDescription` FROM iam_roles";
 
     if (!sSearchWords.empty())
     {
         sSearchWords = '%' + sSearchWords + '%';
-        sSqlQuery+=" WHERE (`roleName` LIKE :SEARCHWORDS OR `roleDescription` LIKE :SEARCHWORDS)";
+        sSqlQuery += " WHERE (`roleName` LIKE :SEARCHWORDS OR `roleDescription` LIKE :SEARCHWORDS)";
     }
 
     if (limit)
-        sSqlQuery+=" LIMIT :LIMIT OFFSET :OFFSET";
+        sSqlQuery += " LIMIT :LIMIT OFFSET :OFFSET";
 
-    sSqlQuery+=";";
-    
+    sSqlQuery += ";";
+
     std::shared_ptr<SQLConnector::QueryInstance> i = _parent->m_sqlConnector->qSelect(sSqlQuery,
-                                          {
-                                              {":SEARCHWORDS",MAKE_VAR(STRING,sSearchWords)},
-                                              {":LIMIT",MAKE_VAR(UINT64,limit)},
-                                              {":OFFSET",MAKE_VAR(UINT64,offset)}
-                                          },
-                                          { &roleName, &description });
+                                                                                      {{":SEARCHWORDS", MAKE_VAR(STRING, sSearchWords)},
+                                                                                       {":LIMIT", MAKE_VAR(UINT64, limit)},
+                                                                                       {":OFFSET", MAKE_VAR(UINT64, offset)}},
+                                                                                      {&roleName, &description});
     while (i->getResultsOK() && i->query->step())
     {
         RoleDetails rDetail;
@@ -158,4 +137,3 @@ std::list<RoleDetails> IdentityManager_DB::Roles_DB::searchRoles(std::string sSe
 
     return ret;
 }
-
